Catch up feature

+++ ulam1/src/ulam/CompilerState.cpp
@@ -62,7 +62,7 @@ namespace MFM {
   static const char * BUILD_DEFAULT_QUARK_FUNCNAME = "getDefaultQuark";
 
   //use of this in the initialization list seems to be okay;
-  CompilerState::CompilerState(): m_programDefST(*this), m_currentFunctionBlockDeclSize(0), m_currentFunctionBlockMaxDepth(0), m_parsingControlLoop(0), m_gotStructuredCommentToken(false), m_parsingConditionalAs(false), m_genCodingConditionalHas(false), m_eventWindow(*this), m_goAgainResolveLoop(false), m_currentSelfSymbolForCodeGen(NULL), m_nextTmpVarNumber(0), m_nextNodeNumber(0)
+  CompilerState::CompilerState(): m_programDefST(*this), m_currentFunctionBlockDeclSize(0), m_currentFunctionBlockMaxDepth(0), m_parsingControlLoop(0), m_gotStructuredCommentToken(false), m_parsingConditionalAs(false), m_genCodingConditionalHas(false), m_eventWindow(*this), m_goAgainResolveLoop(false), m_pendingArgStubContext(0), m_currentSelfSymbolForCodeGen(NULL), m_nextTmpVarNumber(0), m_nextNodeNumber(0)
   {
     m_err.init(this, debugOn, infoOn, warnOn, NULL);
     Token::initTokenMap(*this);
@@ -237,7 +237,7 @@ namespace MFM {
 
   //convenience method (refactors code originally from installSymbol)
   //if exists, just returns it, o.w. makes it; trick to know the base ULAMTYPE
-  UTI CompilerState::makeUlamType(Token typeTok, s32 bitsize, s32 arraysize, UTI classinstanceidx)
+  UTI CompilerState::makeUlamType(Token typeTok, s32 bitsize, s32 arraysize, UTI classinstanceidx, ALT reftype)
   {
     //type names begin with capital letter..and the rest can be either
     u32 typeNameId = getTokenAsATypeNameId(typeTok); //Foo, Int, etc
@@ -253,7 +253,7 @@ namespace MFM {
     if(bitsize == 0)
       bitsize = ULAMTYPE_DEFAULTBITSIZE[bUT];
 
-    UlamKeyTypeSignature key(typeNameId,bitsize,arraysize,classinstanceidx);
+    UlamKeyTypeSignature key(typeNameId, bitsize, arraysize, classinstanceidx, reftype);
 
     UTI uti = Nav;
     UlamType * ut = NULL; //for isDefined.
@@ -286,6 +286,10 @@ namespace MFM {
 	      {
 		//keep classinstanceid of scalar in key
 		assert(suti > 0 && !isComplete(suti));
+	      }
+	    else if(key.getUlamKeyTypeSignatureReferenceType() != ALT_NOT) //array type
+	      {
+		//need a new UTI for reference
 	      }
 	    else
 	      {
@@ -489,12 +493,15 @@ namespace MFM {
     if(findaUTIAlias(auti, mappedUTI))
        return mappedUTI; //anonymous UTI
 
-    //move this test after looking for the mapped class symbol type in "cuti" (always compileThis?)
     UlamType * aut = getUlamTypeByIndex(auti);
+    UlamKeyTypeSignature akey = getUlamKeyTypeSignatureByIndex(auti);
+
+    //if(aut->isReference()) not a special case
+
+    //move this test after looking for the mapped class symbol type in "cuti" (always compileThis?)
     ULAMTYPE bUT = aut->getUlamTypeEnum();
     if(bUT == Class)
       {
-	UlamKeyTypeSignature akey = getUlamKeyTypeSignatureByIndex(auti);
 	SymbolClassName * cnsymOfIncomplete = NULL; //could be a different class than being compiled
 	AssertBool isDefined = alreadyDefinedSymbolClassName(akey.getUlamKeyTypeSignatureNameId(), cnsymOfIncomplete);
 	assert(isDefined);
@@ -506,13 +513,6 @@ namespace MFM {
 	    return (cnsymOfIncomplete->hasMappedUTI(auti, mappedUTI));
 	  }
       }
-
-    // try in Symbol Class..
-    // check for inheritance
-    //    UTI superuti = isClassASubclass(cuti);
-    //if(superuti != Nav)
-    //  return mappedIncompleteUTI(superuti, auti, mappedUTI);
-
     return false; //for compiler
   } //mappedIncompleteUTI
 
@@ -528,6 +528,7 @@ namespace MFM {
     SymbolClassNameTemplate * cnsym = NULL;
     AssertBool isDefined = alreadyDefinedSymbolClassNameTemplate(getCompileThisId(), cnsym);
     assert(isDefined);
+
     UTI mappedUTI;
     if(cnsym->hasInstanceMappedUTI(getCompileThisIdx(), suti, mappedUTI))
       return mappedUTI;  //e.g. decl list
@@ -543,8 +544,23 @@ namespace MFM {
       {
 	AssertBool isDefined = alreadyDefinedSymbolClassName(skey.getUlamKeyTypeSignatureNameId(), cnsymOfIncomplete);
 	assert(isDefined);
+
 	if(!cnsymOfIncomplete->isClassTemplate())
 	  return suti;
+	if(skey.getUlamKeyTypeSignatureReferenceType() == ALT_AS)
+	  {
+	    //UTI asref = getUlamTypeAsRef(getCompileThisIdx(), ALT_AS); //?
+	    //	    return asref;
+	    UTI asref = mapIncompleteUTIForCurrentClassInstance(getUlamTypeAsDeref(suti));
+	    return getUlamTypeAsRef(asref, ALT_AS);
+	  }
+	if(skey.getUlamKeyTypeSignatureReferenceType() == ALT_REF)
+	  {
+	    //UTI asref = getUlamTypeAsRef(getCompileThisIdx(), ALT_REF); //?
+	    //return asref;
+	    UTI asref = mapIncompleteUTIForCurrentClassInstance(getUlamTypeAsDeref(suti));
+	    return getUlamTypeAsRef(asref, ALT_AS);
+	  }
 	if(!((SymbolClassNameTemplate *) cnsymOfIncomplete)->pendingClassArgumentsForStubClassInstance(suti))
 	  return suti;
       }
@@ -589,6 +605,9 @@ namespace MFM {
 
   void CompilerState::mapTypesInCurrentClass(UTI fm, UTI to)
   {
+    if(getReferenceType(to) != ALT_NOT)
+      return; //only a reference to a class type, skip
+
     SymbolClassName * cnsym = NULL;
     AssertBool isDefined = alreadyDefinedSymbolClassName(getCompileThisId(), cnsym);
     assert(isDefined);
@@ -745,15 +764,69 @@ namespace MFM {
 
     if(bUT == Class)
       {
-	return keyOfArg.getUlamKeyTypeSignatureClassInstanceIdx();
+	//return keyOfArg.getUlamKeyTypeSignatureClassInstanceIdx(); // what-if a ref?
+	UTI cuti = keyOfArg.getUlamKeyTypeSignatureClassInstanceIdx(); // what-if a ref?
+	ALT argreftype = keyOfArg.getUlamKeyTypeSignatureReferenceType();
+	ALT creftype = getUlamTypeByIndex(cuti)->getReferenceType();
+	if(argreftype == creftype)
+	  return cuti;
+	// otherwise make a new one as non-classes.
       }
 
     u32 bitsize = keyOfArg.getUlamKeyTypeSignatureBitSize();
-    UlamKeyTypeSignature baseKey(keyOfArg.m_typeNameId, bitsize);  //default array size is zero
+    UlamKeyTypeSignature baseKey(keyOfArg.m_typeNameId, bitsize, NONARRAYSIZE, keyOfArg.m_classInstanceIdx, keyOfArg.m_referenceType);  //default array size is NONARRAYSIZE
 
     UTI buti = makeUlamType(baseKey, bUT); //could be a new one, oops.
     return buti;
   } //getUlamTypeAsScalar
+
+  UTI CompilerState::getUlamTypeAsDeref(UTI utArg)
+  {
+    UlamType * ut = getUlamTypeByIndex(utArg);
+    if(!ut->isReference())
+      return utArg;
+
+    ULAMTYPE bUT = ut->getUlamTypeEnum();
+    UlamKeyTypeSignature keyOfArg = ut->getUlamKeyTypeSignature();
+    u32 bitsize = keyOfArg.getUlamKeyTypeSignatureBitSize();
+    u32 arraysize = keyOfArg.getUlamKeyTypeSignatureArraySize();
+    UTI classidx = keyOfArg.getUlamKeyTypeSignatureClassInstanceIdx();
+
+    if(bUT == Class)
+      return classidx;
+
+    UlamKeyTypeSignature baseKey(keyOfArg.m_typeNameId, bitsize, arraysize, classidx, ALT_NOT);  //default array size is zero
+
+    UTI buti = makeUlamType(baseKey, bUT); //could be a new one, oops.
+    return buti;
+  } //getUlamTypeAsDeref
+
+  UTI CompilerState::getUlamTypeAsRef(UTI utArg)
+  {
+    return getUlamTypeAsRef(utArg, ALT_REF);
+  } //getUlamTypeAsRef
+
+  UTI CompilerState::getUlamTypeAsRef(UTI utArg, ALT altArg)
+  {
+    UlamType * ut = getUlamTypeByIndex(utArg);
+    if(ut->getReferenceType() == altArg)
+      return utArg;
+
+    ULAMTYPE bUT = ut->getUlamTypeEnum();
+    UlamKeyTypeSignature keyOfArg = ut->getUlamKeyTypeSignature();
+
+    u32 bitsize = keyOfArg.getUlamKeyTypeSignatureBitSize();
+    u32 arraysize = keyOfArg.getUlamKeyTypeSignatureArraySize();
+    UTI classidx = keyOfArg.getUlamKeyTypeSignatureClassInstanceIdx();
+    u32 nameid = keyOfArg.m_typeNameId;
+    //    if(classidx != Nav)
+    //  classidx = utArg; //use instance, rather than template in reference key
+    UlamKeyTypeSignature baseKey(nameid, bitsize, arraysize, classidx, altArg);  //default array size is zero
+
+    UTI buti = makeUlamType(baseKey, bUT); //could be a new one, oops.
+
+    return buti;
+  } //getUlamTypeAsRef
 
   UTI CompilerState::getUlamTypeOfConstant(ULAMTYPE etype)
   {
@@ -802,6 +875,12 @@ namespace MFM {
   {
     UlamType * ut = getUlamTypeByIndex(utArg);
     return (ut->getBitSize());
+  }
+
+  ALT CompilerState::getReferenceType(UTI utArg)
+  {
+    UlamType * ut = getUlamTypeByIndex(utArg);
+    return ut->getReferenceType();
   }
 
   bool CompilerState::isComplete(UTI utArg)
@@ -869,8 +948,7 @@ namespace MFM {
     UlamKeyTypeSignature key = ut->getUlamKeyTypeSignature();
 
     //continue with valid number of bits for Class UlamTypes only
-    UlamKeyTypeSignature newkey = UlamKeyTypeSignature(key.getUlamKeyTypeSignatureNameId(), bitsize, arraysize);
-    newkey.append(key.getUlamKeyTypeSignatureClassInstanceIdx());
+    UlamKeyTypeSignature newkey = UlamKeyTypeSignature(key.getUlamKeyTypeSignatureNameId(), bitsize, arraysize, key.getUlamKeyTypeSignatureClassInstanceIdx(), key.getUlamKeyTypeSignatureReferenceType());
 
     if(key == newkey)
       return;
@@ -908,6 +986,9 @@ namespace MFM {
     //bitsize of old could still be "unknown" (before size set, but args known and match 'cuti').
     assert(key1.getUlamKeyTypeSignatureNameId() == key2.getUlamKeyTypeSignatureNameId());
 
+    if(key1.getUlamKeyTypeSignatureReferenceType() != key2.getUlamKeyTypeSignatureReferenceType())
+      return; //don't destory olduti if either is a reference of the other
+
     //removes old key and its ulamtype from map, if no longer pointed to
     deleteUlamKeyTypeSignature(key1, olduti);
     m_indexToUlamKey[olduti] = key2;
@@ -919,6 +1000,49 @@ namespace MFM {
       MSG2(getFullLocationAsString(m_locOfNextLineText).c_str(), msg.str().c_str(), DEBUG);
     }
   } //mergeClassUTI
+
+  //(unused)
+  void CompilerState::rekeyToReferenceUTI(ALT autoreftype, UTI auti)
+  {
+    UlamType * aut = getUlamTypeByIndex(auti);
+
+    UlamKeyTypeSignature key1 = aut->getUlamKeyTypeSignature();
+    UlamKeyTypeSignature newkey = UlamKeyTypeSignature(key1.getUlamKeyTypeSignatureNameId(), key1.getUlamKeyTypeSignatureBitSize(), key1.getUlamKeyTypeSignatureArraySize(), key1.getUlamKeyTypeSignatureClassInstanceIdx(), autoreftype);
+    //UlamKeyTypeSignature newkey = UlamKeyTypeSignature(key1.getUlamKeyTypeSignatureNameId(), key1.getUlamKeyTypeSignatureBitSize(), key1.getUlamKeyTypeSignatureArraySize(), auti, autoreftype);
+
+    if(key1 == newkey)
+      return;
+
+    ULAMTYPE etype = aut->getUlamTypeEnum();
+    bool isAClass = (etype == Class);
+    bool isCustomArray = (isAClass ? aut->isCustomArray() : false);
+    ULAMCLASSTYPE classtype = aut->getUlamClass(); // or notaclass
+
+    //removes old key and its ulamtype from map, if no longer pointed to
+    deleteUlamKeyTypeSignature(key1, auti);
+
+    UlamType * newut = NULL;
+    if(!isDefined(newkey, newut))
+      {
+	newut = createUlamType(newkey, etype);
+	m_definedUlamTypes.insert(std::pair<UlamKeyTypeSignature, UlamType *>(newkey,newut));
+	if(isAClass)
+	  {
+	    ((UlamTypeClass *) newut)->setUlamClass(classtype); //restore from original ut
+	    if(isCustomArray)
+	      ((UlamTypeClass *) newut)->setCustomArray();
+	  }
+      }
+
+    m_indexToUlamKey[auti] = newkey;
+    incrementKeyToAnyUTICounter(newkey, auti);
+    {
+      std::ostringstream msg;
+      msg << "Reference key for (UTI" << auti << ") ";
+      msg << getUlamTypeNameBriefByIndex(auti).c_str();
+      MSG2(getFullLocationAsString(m_locOfNextLineText).c_str(), msg.str().c_str(), DEBUG);
+    }
+  } //rekeyToReferenceUTI
 
   bool CompilerState::isARootUTI(UTI auti)
   {
@@ -1013,7 +1137,7 @@ namespace MFM {
       }
 
     //continue with valid number of bits
-    UlamKeyTypeSignature newkey = UlamKeyTypeSignature(key.getUlamKeyTypeSignatureNameId(), bitsize, arraysize);
+    UlamKeyTypeSignature newkey = UlamKeyTypeSignature(key.getUlamKeyTypeSignatureNameId(), bitsize, arraysize, key.getUlamKeyTypeSignatureClassInstanceIdx(), key.getUlamKeyTypeSignatureReferenceType());
 
     if(key == newkey)
       return;
@@ -1064,8 +1188,12 @@ namespace MFM {
     if(uti == Void)
       return 0;
 
-    s32 arraysize = getArraySize(uti);
-    PACKFIT packed = determinePackable(uti);
+    UlamType * ut = getUlamTypeByIndex(uti);
+    if(ut->isReference())
+      return 1;
+
+    s32 arraysize = ut->getArraySize();
+    PACKFIT packed = ut->getPackable();
 
     if(WritePacked(packed))
       arraysize = 1;
@@ -1096,6 +1224,20 @@ namespace MFM {
       }
     return Nav; //even for non-classes
   } //isClassASubclass
+
+  void CompilerState::resetClassSuperclass(UTI cuti, UTI superuti)
+  {
+    UTI subuti = getUlamTypeAsScalar(cuti); //in case of array
+
+    SymbolClass * csym = NULL;
+    if(alreadyDefinedSymbolClass(subuti, csym))
+      {
+	SymbolClassName * cnsym = NULL;
+	AssertBool isDefined = alreadyDefinedSymbolClassName(csym->getId(), cnsym);
+	assert(isDefined);
+	cnsym->setSuperClassForClassInstance(superuti, cuti);
+      }
+  } //resetClassSuperclass
 
   //return true if the second arg is a superclass of the first arg.
   // i.e. cuti is a subclass of superp. recurses the family tree.
@@ -1194,6 +1336,9 @@ namespace MFM {
 	std::ostringstream msg;
 	msg << "Class without parameters already exists with the same name: ";
 	msg << m_pool.getDataAsString(symptr->getId()).c_str();
+	//msg << " (";
+	//msg << getUlamTypeNameByIndex(symptr->getUlamTypeIdx()).c_str();
+	//msg << ")";
 	MSG2(getFullLocationAsString(m_locOfNextLineText).c_str(), msg.str().c_str(), ERR); //parsing
       }
     return (rtnb && symptr->isClassTemplate());
@@ -1207,6 +1352,9 @@ namespace MFM {
     UTI scalarUTI = uti;
     if(!ut->isScalar())
       scalarUTI = getUlamTypeAsScalar(uti);
+
+    if(ut->isReference())
+      scalarUTI = getUlamTypeAsDeref(scalarUTI); //and deref
 
     SymbolClassName * cnsym = NULL;
     if(alreadyDefinedSymbolClassName(ut->getUlamKeyTypeSignature().getUlamKeyTypeSignatureNameId(), cnsym))
@@ -1610,7 +1758,7 @@ bool CompilerState::isFuncIdInAClassScope(UTI cuti, u32 dataindex, Symbol * & sy
 
     if(m_currentFunctionReturnNodes.empty())
       {
-	if(it != Void && !fsym->isNativeFunctionDeclaration() && !fsym->isPureVirtualFunction())
+	if(it != Void && !fsym->isNativeFunctionDeclaration() && (!fsym->isVirtualFunction() || !fsym->isPureVirtualFunction()))
 	  {
 	    std::ostringstream msg;
 	    msg << "Function '" << m_pool.getDataAsString(fsym->getId()).c_str();
@@ -1636,6 +1784,11 @@ bool CompilerState::isFuncIdInAClassScope(UTI cuti, u32 dataindex, Symbol * & sy
 	    m_err.buildMessage(rNode->getNodeLocationAsString().c_str(), msg.str().c_str(), "MFM::NodeReturnStatement", "checkAndLabelType", rNode->getNodeLocation().getLineNo(), MSG_DEBUG);
 	    continue;
 	  }
+
+	//treat ALT_AS as defref'd type; ALT_REF is its own type.
+	//ALT reftype = getUlamTypeByIndex(rType)->getReferenceType();
+	//if(reftype == ALT_AS)
+	// rType = getUlamTypeAsDeref(rType);
 
 	if(UlamType::compare(rType, it, *this) != UTIC_SAME)
 	  {
@@ -2096,6 +2249,7 @@ bool CompilerState::isFuncIdInAClassScope(UTI cuti, u32 dataindex, Symbol * & sy
     assert(isQuark);
     NodeBlockClass * classNode = csym->getClassBlockNode();
     assert(classNode);
+
     NodeBlockFunctionDefinition * func = classNode->findToIntFunctionNode();
     return (func != NULL);
   } //quarkHasAToIntMethod
@@ -2321,7 +2475,24 @@ bool CompilerState::isFuncIdInAClassScope(UTI cuti, u32 dataindex, Symbol * & sy
     SymbolClassName * cnsym = NULL;
     AssertBool isDefined = alreadyDefinedSymbolClassName(getCompileThisId(), cnsym);
     assert(isDefined);
-    return cnsym->findNodeNoInAClassInstance(getCompileThisIdx(), n);
+    Node * rtnNode = cnsym->findNodeNoInAClassInstance(getCompileThisIdx(), n);
+
+    //last resort, if we are in the middle of resolving pending args for a stub
+    // and to do constant folding, we need to find the parent node that's in the
+    // stub's resolver, NOT the context where the stub appears.
+    if(!rtnNode)
+      {
+	UTI stubuti = m_pendingArgStubContext;
+	if( stubuti != Nav)
+	  {
+	    u32 stubid = getUlamKeyTypeSignatureByIndex(stubuti).getUlamKeyTypeSignatureNameId();
+	    SymbolClassNameTemplate * cntsym = NULL;
+	    AssertBool isDefined = alreadyDefinedSymbolClassNameTemplate(stubid, cntsym);
+	    assert(isDefined);
+	    rtnNode =  cntsym->findNodeNoInAClassInstance(stubuti, n);
+	  }
+      }
+    return rtnNode;
   } //findNodeNoInThisClass
 
   Node * CompilerState::findNodeNoInAClass(NNO n, UTI cuti)
@@ -2373,14 +2544,6 @@ bool CompilerState::isFuncIdInAClassScope(UTI cuti, u32 dataindex, Symbol * & sy
   SymbolClass * CompilerState::getCurrentSelfSymbolForCodeGen()
   {
     return (SymbolClass *) m_currentSelfSymbolForCodeGen;
-#if 0
-    // no longer all self in h/as conditionals.
-    Symbol * selfsym = NULL;
-    u32 selfid = m_pool.getIndexForDataString("self");
-    bool isDefined = alreadyDefinedSymbol(selfid, selfsym, hazyKin);
-    assert(isDefined);
-    return (SymbolClass *) selfsym;
-#endif
   } //getCurrentSelfSymbolForCodeGen
 
   NodeBlock * CompilerState::getCurrentBlock()
